M2RS-34 refactor: Data 관련 Exception 에 대한 상위 Exception 추가
