+++ m2rs-core/src/main/java/com/m2rs/core/commons/exception/DataException.java
@@ -1,0 +1,9 @@
+package com.m2rs.core.commons.exception;
+
+public class DataException extends ServiceRuntimeException {
+
+    public DataException(String message) {
+        super(message);
+    }
+
+}
