+++ m2rs-core/src/main/java/com/m2rs/core/commons/exception/data/AlreadyExistDataException.java
@@ -1,6 +1,9 @@
-package com.m2rs.core.commons.exception;
+package com.m2rs.core.commons.exception.data;
 
-public class AlreadyExistDataException extends ServiceRuntimeException {
+import com.m2rs.core.commons.exception.DataException;
+import com.m2rs.core.commons.exception.ServiceRuntimeException;
+
+public class AlreadyExistDataException extends DataException {
 
     public AlreadyExistDataException(Class<?> target, String data) {
         super(String.format("'%s' already exist in the '%s'.", data, target.getSimpleName()));
