+++ m2rs-user-service/src/main/java/com/m2rs/userservice/controller/GlobalControllerAdvice.java
@@ -1,6 +1,8 @@
 package com.m2rs.userservice.controller;
 
-import com.m2rs.core.commons.exception.DataReferenceException;
+import com.m2rs.core.commons.exception.DataException;
+import com.m2rs.core.commons.exception.data.AlreadyExistDataException;
+import com.m2rs.core.commons.exception.data.DataReferenceException;
 import lombok.extern.slf4j.Slf4j;
 
 import org.springframework.http.HttpStatus;
@@ -18,7 +20,7 @@ import static com.m2rs.core.commons.model.api.response.ApiResult.error;
 public class GlobalControllerAdvice {
 
     @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler({IllegalArgumentException.class, DataReferenceException.class})
+    @ExceptionHandler({IllegalArgumentException.class, DataException.class})
     public <T> ApiResult<T> badRequest(Exception e) {
         log.warn(e.getMessage());
